+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Changed
-- Using the new script system by default
+- Using the new script system by default, you can still switch to the old one using `-oldss` command line parameter
 - Kill log now shows RunOver icon when killing players with object
 - Improve First Person spectator vehicle view
 - Remove transparency for enemy in radar